# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BudgetSpec": {
        "properties": {
          "content_purchase_cost": {
            "const": 0,
            "default": 0,
            "title": "Content Purchase Cost",
            "type": "integer"
          },
          "daily_requests": {
            "default": 96,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Daily Requests",
            "type": "integer"
          }
        },
        "title": "BudgetSpec",
        "type": "object"
      },
...
      "ErrorView": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "request_id"
        ],
        "title": "ErrorView",
        "type": "object"
      },
...
      "MonitorInput": {
        "additionalProperties": false,
        "properties": {
          "budget": {
            "$ref": "#/components/schemas/BudgetSpec"
          },
          "query_spec": {
            "$ref": "#/components/schemas/QuerySpec"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleSpec"
          },
          "source_ids": {
            "items": {
              "enum": [
                "x",
                "bilibili",
                "weibo",
                "xiaohongshu",
                "douyin",
                "bluesky"
              ],
              "type": "string"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Source Ids",
            "type": "array"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "query_spec",
          "source_ids"
        ],
        "title": "MonitorInput",
        "type": "object"
      },
      "MonitorPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MonitorView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "MonitorPage",
        "type": "object"
      },
      "MonitorStateChange": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "MonitorStateChange",
        "type": "object"
      },
      "MonitorUpdate": {
        "additionalProperties": false,
        "properties": {
          "budget": {
            "$ref": "#/components/schemas/BudgetSpec"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "query_spec": {
            "$ref": "#/components/schemas/QuerySpec"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleSpec"
          },
          "source_ids": {
            "items": {
              "enum": [
                "x",
                "bilibili",
                "weibo",
                "xiaohongshu",
                "douyin",
                "bluesky"
              ],
              "type": "string"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Source Ids",
            "type": "array"
          },
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "query_spec",
          "source_ids",
          "expected_version"
        ],
        "title": "MonitorUpdate",
        "type": "object"
      },
      "MonitorView": {
        "properties": {
          "budget": {
            "$ref": "#/components/schemas/BudgetSpec"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "query_spec": {
            "$ref": "#/components/schemas/QuerySpec"
          },
          "schedule": {
            "$ref": "#/components/schemas/ScheduleSpec"
          },
          "source_ids": {
            "items": {
              "enum": [
                "x",
                "bilibili",
                "weibo",
                "xiaohongshu",
                "douyin",
                "bluesky"
              ],
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "state": {
            "enum": [
              "draft",
              "active",
              "paused"
            ],
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "state",
          "current_version",
          "query_spec",
          "source_ids",
          "schedule",
          "budget",
          "created_at",
          "updated_at"
        ],
        "title": "MonitorView",
        "type": "object"
      },
...
      "QuerySpec": {
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Aliases",
            "type": "array"
          },
          "exclude": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Exclude",
            "type": "array"
          },
          "include_all": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Include All",
            "type": "array"
          },
          "include_any": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Include Any",
            "type": "array"
          }
        },
        "required": [
          "include_any"
        ],
        "title": "QuerySpec",
        "type": "object"
      },
      "ScheduleSpec": {
        "properties": {
          "interval_minutes": {
            "default": 60,
            "maximum": 1440.0,
            "minimum": 15.0,
            "title": "Interval Minutes",
            "type": "integer"
          },
          "retention_days": {
            "default": 7,
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "type": "integer"
          }
        },
        "title": "ScheduleSpec",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OwnerSession": {
        "in": "cookie",
        "name": "hk_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-owner learning workspace. Unsafe requests require an exact allowed Origin; authenticated writes also require a session-bound X-CSRF-Token. Monitor configuration history and zero-network query previews are available; collection remains disabled until a source passes rights and pipeline admission.",
    "title": "HotKey API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/monitors": {
      "get": {
        "operationId": "listMonitors",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Monitors",
        "tags": [
          "monitoring"
        ]
      },
      "post": {
        "operationId": "createMonitor",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Create Monitor",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/monitors/{identity}": {
      "patch": {
        "operationId": "updateMonitor",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "State conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Update Monitor",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/monitors/{identity}/activate": {
      "post": {
        "operationId": "activateMonitor",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorStateChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "State conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Activate Monitor",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/monitors/{identity}/pause": {
      "post": {
        "operationId": "pauseMonitor",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorStateChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "State conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Request content too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorView"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "OwnerSession": []
          }
        ],
        "summary": "Pause Monitor",
        "tags": [
          "monitoring"
        ]
      }
    },
...
  }
}